+++ state/export_test.go
@@ -55,3 +55,11 @@ func (p *Pruner) FindMinRetainHeight() int64 {
 func (p *Pruner) PruneABCIResToRetainHeight(lastRetainHeight int64) int64 {
 	return p.pruneABCIResToRetainHeight(lastRetainHeight)
 }
+
+func Int64ToBytes(val int64) []byte {
+	return int64ToBytes(val)
+}
+
+func Int64FromBytes(val []byte) int64 {
+	return int64FromBytes(val)
+}